feat: enhance user system and improve overall experience

+++ backend/src/routes/vocabulary.rs
@@ -36,7 +36,7 @@ pub struct SaveVocabularyRequest {
 
 #[derive(Serialize)]
 pub struct SaveVocabularyResponse {
-    id: i64,
+    id: i32,
 }
 
 async fn save_vocabulary(
